requirements.txt
langchain 
faiss-cpu
arxiv 
pypdf 
tiktoken 
fastapi 
uvicorn 
sentence-transformers 
google-generativeai
python-dotenv